Edits

+++ server/scm/Git.py
@@ -1,15 +1,22 @@
 __author__ = 'mohamed'
 from .BaseSCM import BaseSCM
 import Common
+import os
 
 class GIT(BaseSCM):
 
     def __init__(self,workdir,repo=""):
         self.workdir=workdir
         self.repo=repo
+    def check_dir(self):
+        if not os.path.exists(self.workdir):
+            return b"ERR: The working dir does not exists , please clone first."
+        return True
     def setKey(self,key):
         from os.path import expanduser
         home = expanduser("~")
+        if not os.path.exists(home):
+            os.makedirs(home+"/.ssh")
         keyfile=home+"/.ssh/id_rsa"
         Common.run("chmod 700 %s"%keyfile)
         f=open(keyfile,"w")
@@ -19,14 +26,26 @@ class GIT(BaseSCM):
     def get_clone_cmd(self):
         return "git clone %s %s"%(self.repo,self.workdir)
     def get_pull_cmd(self):
+        x = self.check_dir()
+        if x != True:
+            return x
         return "cd %s; git pull"%self.workdir
     def get_list_tags_cmd(self):
+        x = self.check_dir()
+        if x != True:
+            return x
         return "cd %s;git tag -l"%self.workdir
 
     def get_list_branches(self):
+        x = self.check_dir()
+        if x != True:
+            return x
         return "cd %s;git branch -a" % self.workdir
 
     def get_switch_to_tag_cmd(self,tag):
+        x = self.check_dir()
+        if x != True:
+            return x
         return "cd %s; git checkout tags/%s"%(self.workdir,tag)
     def get_history_cmd(self,options={},limit=0):
         branch=options.get("branch","")
@@ -37,10 +56,16 @@ class GIT(BaseSCM):
         cmd += ' --pretty=format:"%H,,%h,,%an,,%ar,,%s,,%cd"  | cat -'
         return cmd
     def switch_to_histroy_cmd(self,commit):
+        x = self.check_dir()
+        if x != True:
+            return x
         return 'cd %s; git reset --hard %s'%(self.workdir,commit)
     def commit_diff_cmd(self,commit):
         Common.run(self.get_pull_cmd())
         return 'git rev-list %s..HEAD'%(commit)
     def get_changelog(self,since,to):
+        x = self.check_dir()
+        if x != True:
+            return x
         return "cd %s;git shortlog %s..%s"%(self.workdir,since,to)
